[DEV] Refactor/structure (#34)
* refactor: Delete the old folder structure

* chore: Update package lock json

* refactor(auth): Move from component/pages to pages root directory

* refactor(dashboard): Move old directory file to new one

* refactor(hooks): Move all hooks into hooks folder in root directory
