--- app/api/axios.ts
@@ -1,18 +1,0 @@
-import axios from 'axios'
-
-const api = axios.create({
-  baseURL: 'https://melancong-be.vercel.app'
-})
-
-api.interceptors.request.use(
-  config => {
-    const token = localStorage.getItem('token')
-    if (token) {
-      config.headers.Authorization = `Bearer ${token}`
-    }
-    return config
-  },
-  error => Promise.reject(error)
-)
-
-export default api
