--- app/api/error.ts
@@ -1,9 +1,0 @@
-interface ErrorResponse {
-  response: {
-    data: {
-      message: string
-    }
-  }
-}
-
-export default ErrorResponse
